[tests] msm has its own subdirectory and added unit test for count matrix cpp impl
