+++ tests/decomposition/test_vamp.py
@@ -25,9 +25,9 @@ import unittest
 import numpy as np
 
 from sktime.data.util import timeshifted_split
-from sktime.decomposition.vamp import VAMP, VAMPModel
+from sktime.decomposition.vamp import VAMP
 from sktime.markov._base import cvsplit_dtrajs
-from tests.markov.test_msm import estimate_markov_model
+from tests.markov.msm.test_msm import estimate_markov_model
 
 
 def random_matrix(n, rank=None, eps=0.01):
