+++ tests/markov/msm/test_bayesian_msm.py
