+++ tests/markov/test_sample.py
@@ -1,37 +1,9 @@
-import time
 import unittest
 
 import numpy as np
-
 import sktime.markov.sample as sample
 
-
-class timing(object):
-    """A timing context manager
-
-    Examples
-    --------
-    >>> long_function = lambda : None
-    >>> with timing('long_function'):
-    ...     long_function()
-    long_function: 0.000 seconds
-    """
-
-    def __init__(self, name='block'):
-        self.name = name
-        self.time = 0
-        self.start = None
-        self.end = None
-
-    def __enter__(self):
-        self.start = time.time()
-        return self
-
-    def __exit__(self, ty, val, tb):
-        self.end = time.time()
-        self.time = self.end - self.start
-        print("%s: %0.3f seconds" % (self.name, self.time))
-        return False
+from tests.util import timing
 
 
 class TestSampleIndices(unittest.TestCase):
